+++ gulpfiles/lib/tasks.js
@@ -1,6 +1,7 @@
+import path from 'path';
 import gulp from 'gulp';
 import gutil from 'gulp-util';
-import babel from 'gulp-babel';
+import gulpBabel from 'gulp-babel';
 import rimraf from 'rimraf';
 import glob from 'glob';
 import eslint from 'eslint';
@@ -8,73 +9,52 @@ import ESDoc from 'esdoc/out/src/ESDoc';
 import ESDocPublisher from 'esdoc/out/src/Publisher/publish';
 import * as $ from './helper';
 
-// -----------------------------------------------
-// Main tasks
-// -----------------------------------------------
+/**
+ * @fileoverview
+ * This file defines the core tasks executed by gulp.
+ * All tasks must be curried to allow clients to pass options.
+ */
 
-gulp.task('default', [
-  'lint:watch',
-  'test:watch'
-]);
-
-gulp.task('check:strict', [
-  '_lint:disallowWarns',
-  'check'
-]);
-
-gulp.task('check', [
-  'lint:all',
-  'test',
-  'doc'
-]);
-
-gulp.task('lint:all', [
-  'lint',
-  'lint:gulp:fix',
-  'depcop'
-]);
-
-gulp.task('lint', [
-  'lint:lib:fix',
-  'lint:test:fix',
-  'lint:bin:fix'
-]);
+/**
+ * The object to store any state about tasks.
+ */
+export const state = {};
 
 // -----------------------------------------------
 // Build tasks
 // -----------------------------------------------
 
-gulp.task('clean', done => {
+export const clean = () => done => {
   rimraf($.GLOB.build, done);
-});
+};
 
-gulp.task('build', ['clean'], () => {
+export const babel = () => () => {
   return gulp.src($.GLOB.lib)
-    .pipe(babel())
+    .pipe(gulpBabel())
     .pipe(gulp.dest($.PATH.build));
-});
+};
 
-gulp.task('watch', ['clean'], () => {
+export const babelWatch = () => () => {
   $.runAndWatch($.GLOB.lib, $.GLOB.lib, path => {
     gutil.log(gutil.colors.cyan('babel:'), path);
     gulp.src(path, { base: $.PATH.lib })
-      .pipe(babel())
+      .pipe(gulpBabel())
       .on('error', ex => console.log(ex.stack))
       .pipe(gulp.dest($.PATH.build));
   });
-});
+};
 
 // -----------------------------------------------
 // Test tasks
 // -----------------------------------------------
 
-gulp.task('test', ['build', 'test:prepare'], done => {
+export const test = () => done => {
   $.runTests($.GLOB.spec).then(err => {
     done(err ? new Error('Mocha tests failed.') : undefined);
   }, done);
-});
+};
 
-gulp.task('test:watch', ['test:prepare'], () => {
+export const testWatch = () => () => {
   function test() {
     $.runTests(
       $.GLOB.spec,
@@ -102,60 +82,42 @@ gulp.task('test:watch', ['test:prepare'], () => {
   });
 
   $.runAndWatch($.GLOB.spec, null, () => test());
-});
+};
 
-gulp.task('test:prepare', () => {
+export const testPrepare = () => () => {
   require('babel-core/register');
-});
+};
 
 // -----------------------------------------------
 // Lint tasks
 // -----------------------------------------------
 
-[
-  'bin', 'lib', 'test', 'gulp'
-]
-.forEach(target => {
-  const lintTaskName = `lint:${target}`;
+export const lint = (target, options = {}) => () => {
+  const { fix } = options;
+  const globPattern = $.GLOB[target];
+  $.lintFiles([globPattern], { fix }, state.disallowWarns);
+};
 
-  // Define a lint task.
-  gulp.task(lintTaskName, () => {
-    const globPattern = $.GLOB[target];
-    $.lintFiles([globPattern]);
-  });
-
-  // Define a lint task which uses autofix feature.
-  gulp.task(`${lintTaskName}:fix`, [
-    '_lint:enableFix',
-    lintTaskName
-  ]);
-});
-
-gulp.task('lint:watch', () => {
+export const lintWatch = targets => () => {
   const linter = new eslint.CLIEngine();
+
   function lintAndReport(path) {
     const report = linter.executeOnFiles([path]);
     const formatter = linter.getFormatter();
     console.log(formatter(report.results));
   }
 
-  $.runAndWatch($.GLOB.lib, $.GLOB.lib, lintAndReport);
-  $.runAndWatch($.GLOB.test, $.GLOB.test, lintAndReport);
-});
+  targets.forEach(globPattern => {
+    $.runAndWatch(globPattern, globPattern, lintAndReport);
+  });
+};
 
-gulp.task('_lint:enableFix', () => {
-  $.lintFiles.fixEnabled = true;
-});
-
-gulp.task('_lint:disallowWarns', () => {
-  $.lintFiles.disallowWarns = true;
-});
-
-gulp.task('depcop', () => {
+export const depcop = () => () => {
 
   // Don't load 'depcop' at the top level
   // because the source code may be unstable.
-  const { makeDepcop } = require('../lib');
+  const libraryPath = path.resolve($.ROOT_DIR, 'build');
+  const { makeDepcop } = require(libraryPath);
 
   const depcop = makeDepcop();
   const result = depcop.runValidations();
@@ -165,13 +127,13 @@ gulp.task('depcop', () => {
     console.log(format(result));
     throw new Error('depcop warned some dependencies');
   }
-});
+};
 
 // -----------------------------------------------
 // Document tasks
 // -----------------------------------------------
 
-gulp.task('doc', () => {
-  const config = require('../esdoc.json');
+export const esdoc = () => () => {
+  const config = require(`${$.ROOT_DIR}/esdoc.json`);
   ESDoc.generate(config, ESDocPublisher);
-});
+};