+++ test/lib/codeAnalyzer/ImportedModuleCollector.spec.js
@@ -33,6 +33,16 @@ describe('ImportedModuleCollector', () => {
         'module5', 'module6'
       ]
     }],
+    ['recognizes \'export\'s that imports a module', {
+      code: `
+        export * from 'foo';
+        export { a, b, c } from 'bar';
+        export { a as A, b as B, c } from 'baz';
+      `,
+      modules: [
+        'foo', 'bar', 'baz'
+      ]
+    }],
     ['recognizes \'require\' calls', {
       code: `
         var foo = require('foo');
@@ -48,10 +58,22 @@ describe('ImportedModuleCollector', () => {
         'qux', 'inner'
       ]
     }],
+    ['ignores \'export\'s that just exports something', {
+      code: `
+      export { foo };
+      export { a as default, b as B };
+      export const module1 = 1;
+      export function hello() {};
+      export default function greet() {};
+      `,
+      modules: []
+    }],
     ['ignores invalid loadings', {
       code: `
         import '';
         import ' ';
+
+        export * from '';
 
         require(variable);
         require('exp' + 'ression');
@@ -71,6 +93,9 @@ describe('ImportedModuleCollector', () => {
         import qux from './some/dir/qux';
         import quux from '../../../quux';
 
+        export * from './some/dir/qux';
+        export { a } from '../../../quux';
+
         require('foo2');
         require('./bar');
         require('../baz');
@@ -88,6 +113,9 @@ describe('ImportedModuleCollector', () => {
         import _bar from 'bar';
         import assert from 'assert';
         import _baz from 'baz';
+
+        export * from 'path';
+        export { equal } from 'assert';
 
         if (shouldLoad) {
           var fs = require('fs');
@@ -111,6 +139,9 @@ describe('ImportedModuleCollector', () => {
         import bar from 'bar/sub/b';
         import baz from 'baz/inner/SomeClass';
 
+        export * from 'qux/sub/a';
+        export * from 'quux/inner/SomeClass';
+
         if (shouldLoad) {
           var foo = require('foo2');
           var bar = require('bar2/sub/a');
@@ -120,6 +151,7 @@ describe('ImportedModuleCollector', () => {
       `,
       modules: [
         'foo', 'bar', 'baz',
+        'qux', 'quux',
         'foo2', 'bar2', 'baz2'
       ]
     }]